# CSE341_Final_Project_Refueled/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Eko-Store API",
    "version": "1.0.0",
    "description": "API documentation for the Eko-Store application"
  },
  "paths": {
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Retrieve a list of users",
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Retrieve a single user by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "A single user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The user ID"
          },
          "name": {
            "type": "string",
            "description": "The user's name"
          },
          "email": {
            "type": "string",
            "description": "The user's email"
          }
        }
      }
    }
  }
}